+++ usecases/Concrete/tests/test_experimental_setups.py
@@ -1,14 +1,14 @@
 import numpy as np
 
-import usecases.Concrete.simulation_model as Simulation
+import fenics_concrete
 
 import pytest
 
 
 @pytest.mark.parametrize("dim", [2, 3])
 @pytest.mark.parametrize("boundary_setting", ['free', 'fixed'])
-def test_experiemental_setup(dim, boundary_setting):
-    parameters = Simulation.Parameters()  # using the current default values
+def test_experiemental_setup_2(dim, boundary_setting):
+    parameters = fenics_concrete.Parameters()  # using the current default values
 
     parameters['mesh_density'] = 4
     parameters['log_level'] = 'WARNING'
@@ -21,9 +21,9 @@ def test_experiemental_setup(dim, boundary_setting):
 
     displacement = -parameters.height * 0.1
 
-    experiment = Simulation.Setups.ConcreteCylinderExperiment(parameters)
+    experiment = fenics_concrete.ConcreteCylinderExperiment(parameters)
 
-    problem = Simulation.Models.LinearElasticity(experiment, parameters)
+    problem = fenics_concrete.LinearElasticity(experiment, parameters)
 
     problem.experiment.apply_displ_load(displacement)
