+++ usecases/Concrete/tests/test_material_model.py
@@ -1,11 +1,11 @@
 import numpy as np
 
-import usecases.Concrete.simulation_model as Simulation
+import fenics_concrete
 
 import pytest
 
 def simple_setup(p, displacement, sensor):
-    parameters = Simulation.Parameters()  # using the current default values
+    parameters = fenics_concrete.Parameters()  # using the current default values
 
     parameters['log_level'] = 'WARNING'
     parameters['bc_setting'] = 'free'
@@ -13,9 +13,9 @@ def simple_setup(p, displacement, sensor):
 
     parameters = parameters + p
 
-    experiment = Simulation.Setups.ConcreteCylinderExperiment(parameters)
+    experiment = fenics_concrete.ConcreteCylinderExperiment(parameters)
 
-    problem = Simulation.Models.LinearElasticity(experiment, parameters)
+    problem = fenics_concrete.LinearElasticity(experiment, parameters)
     problem.add_sensor(sensor)
 
     problem.experiment.apply_displ_load(displacement)
@@ -27,7 +27,7 @@ def simple_setup(p, displacement, sensor):
 
 # testing the linear elastic response
 def test_force_response_2D():
-    p = Simulation.Parameters()  # using the current default values
+    p = fenics_concrete.Parameters()  # using the current default values
 
     p['E'] = 1023
     p['nu'] = 0.0
@@ -36,13 +36,13 @@ def test_force_response_2D():
     displacement = -3
     p['dim'] = 2
 
-    sensor = Simulation.sensors.ReactionForceSensor()
+    sensor = fenics_concrete.sensors.ReactionForceSensor()
     measured = simple_setup(p, displacement, sensor)
 
     assert measured == pytest.approx(p.E*p.radius*2*displacement/p.height)
 
 def test_force_response_3D():
-    p = Simulation.Parameters()  # using the current default values
+    p = fenics_concrete.Parameters()  # using the current default values
 
     p['E'] = 1023
     p['nu'] = 0.0
@@ -51,7 +51,7 @@ def test_force_response_3D():
     displacement = -3
     p['dim'] = 3
 
-    sensor = Simulation.sensors.ReactionForceSensor()
+    sensor = fenics_concrete.sensors.ReactionForceSensor()
     measured = simple_setup(p, displacement, sensor)
 
     # due to meshing errors, only aprroximate results to be expected. within 1% is good enough
